==> 00. Check that we have all the files and get the paths.ipynb <==
## Step 1: find all the filenames that we should have in the archive

import pandas as pd

df_navigation = pd.read_csv('ArtDATIS_navigation_table.tsv',sep='\t')

df_navigation.head()

df_files_and_descriptions = df_navigation[['part_of.title','images.reference']].dropna()

df_files_and_descriptions['filename'] = df_files_and_descriptions.apply(lambda row: row['images.reference'].split('\\')[-1].split('.')[0],
                                                                                                               axis=1
                                                                    )

df_files_and_descriptions

filenames_from_rkd = list(df_files_and_descriptions['filename'])

## Step 2: check all folders that contain archive files, get a list of filenames

import os
def getListOfFiles(dirName):
    # create a list of file and sub directories 
    # names in the given directory 
    listOfFile = os.listdir(dirName)
    allFiles = list()
    # Iterate over all the entries
    for entry in listOfFile:
        # Create full path
        fullPath = os.path.join(dirName, entry)
        # If entry is a directory then get the list of files in this directory 
        if os.path.isdir(fullPath):
            allFiles = allFiles + getListOfFiles(fullPath)
        else:
            allFiles.append(fullPath)
                
    return allFiles

all_files_old = getListOfFiles('/ivi/ilps/projects/ArtDATIS/valkema/')

all_files_new = getListOfFiles('/ivi/ilps/personal/vprovat/missing_folders_ArtDATIS')

good_files = [file for file in all_files_old+all_files_new if file.endswith('.tif')]

good_file_names = [name.split('/')[-1].split('.')[0] for name in good_files]

len(good_files)

good_files[:5]

missing_files = list(set(filenames_from_rkd).difference(set(good_file_names)))

# todo: filter the table by words "drawing", "photo", ...

len(missing_files)

df_files_and_descriptions['is_missing'] = df_files_and_descriptions.apply(lambda row: row['filename'] in missing_files,
                                                       axis=1)

paths_dict = dict([(name, path) for name, path in zip(good_file_names, good_files)])

for filename in missing_files:
    paths_dict[filename] = None

df_files_and_descriptions['path_to_image'] = df_files_and_descriptions.apply(lambda row: paths_dict[row['filename']],
                                                       axis=1)

df_files_and_descriptions.to_csv('archive_files_with_paths_and_descriptions.tsv',sep='\t', quoting=csv.QUOTE_NONNUMERIC)

df_files_and_descriptions

df_files_and_descriptions[df_files_and_descriptions['is_missing']].to_csv('missing_files.tsv',sep='\t', quoting=csv.QUOTE_NONNUMERIC)

names = sorted(set([name for name in df_navigation['part_of.title'].tolist() if isinstance(name,str)]))
names

corr = [name for name in names if name.lower().find('correspondence')!=-1 or name.lower().find('letter')!=-1]

corr

len(corr)
