==> 01. Classify typed vs non-typed.ipynb <==
def extract_filename(path):
    return path.split('/')[-1].split('.')[0].split('_')[0]

import os
def getListOfFiles(dirName):
    # create a list of file and sub directories 
    # names in the given directory 
    listOfFile = os.listdir(dirName)
    allFiles = list()
    # Iterate over all the entries
    for entry in listOfFile:
        # Create full path
        fullPath = os.path.join(dirName, entry)
        # If entry is a directory then get the list of files in this directory 
        if os.path.isdir(fullPath):
            allFiles = allFiles + getListOfFiles(fullPath)
        else:
            allFiles.append(fullPath)
                
    return allFiles

import pandas as pd


df_files = pd.read_csv('archive_files_with_paths_and_descriptions.tsv',sep='\t')

def recover_filename(f_raw):
    f = str(f_raw)
    while len(f) < len('0001246757'):
        f = '0'+f
    return f

df_files['filename'] = df_files.apply(lambda row: recover_filename(row['filename']),
                                       axis=1)

df_files.head()

# todo: classify typed vs handwritten (or get info for those already classified)

path_typed_old = '/ivi/ilps/projects/ArtDATIS/artdatis/tagging/OCRed/typed/'
path_typed_new = '/ivi/ilps/personal/vprovat/ocr/newest_results/typed/'

typed_files = list(set(getListOfFiles(path_typed_old)+getListOfFiles(path_typed_new)))

texts_typed = [f for f in typed_files if f.endswith('text.txt')]

len(texts_typed)

# filenames_typed = set([f.split('/')[-1].split('.')[0] for f in texts_typed])

path_handwritten = '/ivi/ilps/personal/vprovat/backup_ocr/handwritten/'
path_no_text = '/ivi/ilps/personal/vprovat/backup_ocr/no_text/'
path_no_text_new = '/ivi/ilps/personal/vprovat/backup_ocr/new_results/non-typed/'


non_typed_file_paths = [f for f in list(set(getListOfFiles(path_handwritten)+getListOfFiles(path_no_text)+getListOfFiles(path_no_text_new)))
                        if f.endswith('path.txt')]

non_typed_file_osds = [f for f in list(set(getListOfFiles(path_handwritten)+getListOfFiles(path_no_text)+getListOfFiles(path_no_text_new)))
                        if f.endswith('osd.txt')]

dict_typed_text_paths = dict([(extract_filename(path), path) for path in texts_typed])
dict_non_typed_paths = dict([(extract_filename(path), path) for path in non_typed_file_paths])

# sanity check
set(dict_typed_text_paths.keys()).intersection(set(dict_non_typed_paths.keys()))

# another sanity check: do we have them all?
len(set(dict_typed_text_paths.keys()))+len(set(dict_non_typed_paths.keys()))-len(df_files)

Now, let's try to detect handwritten files among the non-typed ones (and label others as likely drawings):

drawing_files = set()
handwritten_files = set()
dict_handwritten_osd_paths = {}

from tqdm import tqdm

for path in tqdm(non_typed_file_osds):
    name = extract_filename(path)
    osd = open(path,'r').read()
    if not osd:
        drawing_files.add(name)
    else:
        handwritten_files.add(name)
        dict_handwritten_osd_paths[name] = path

typed_files = set(dict_typed_text_paths.keys())

# Handling empty stuff to avoid errors
for f in df_files['filename'].tolist():
    if f not in dict_typed_text_paths:
        dict_typed_text_paths[f] = None
    if f not in dict_handwritten_osd_paths:
        dict_handwritten_osd_paths[f] = None

def get_page_class(f):
    if f in typed_files:
        return "typed"
    if f in handwritten_files:
        return "likely handwritten"
    if f in drawing_files:
        return "likely drawing"
    return None

df_files['page_type'] = df_files.apply(lambda row: get_page_class(row['filename']),
                                       axis=1)

df_files['path_to_text'] = df_files.apply(lambda row: dict_typed_text_paths[row['filename']],
                                       axis=1)

df_files['path_to_handwritten_info'] = df_files.apply(lambda row: dict_handwritten_osd_paths[row['filename']],
                                       axis=1)

df_files.head()

remaining_paths_to_classify = df_files[df_files['page_type'].isna()]['path_to_image'].tolist()

len(remaining_paths_to_classify)

# Classify the missing files with tesseract

import pytesseract
import json

try:
    from PIL import Image
except ImportError:
    import Image

RES_PATH = '/ivi/ilps/personal/vprovat/OCR_for_missing_folders'
    
    
def process_document(path):
    name = extract_filename(path)
    try:
        img = Image.open(path)
    except: # path not found
        return
    
    try:
        osd_dict = pytesseract.image_to_osd(img, output_type='dict')
    except: # Too few characters for osd -> most likely blank or drawing, could also be handwritten
        with open(f"{RES_PATH}/likely_drawing/{name}_path.txt","w") as f:
            f.write(path)
        return
        
    
    if osd_dict['script'] == 'Latin': # Is a typed document
        with open(f"{RES_PATH}/typed/{name}_path.txt","w") as f:
            f.write(path)
            
        text = pytesseract.image_to_string(img)
        with open(f"{RES_PATH}/typed/{name}_text.txt","w") as f:
            f.write(text)
    else: # We think it's handwritten, saving osd just in case
        with open(f"{RES_PATH}/likely_handwritten/{name}_path.txt","w") as f:
            f.write(path)
        with open(f"{RES_PATH}/likely_handwritten/{name}_osd.txt","w") as f:
            f.write(json.dumps(osd_dict))

# import parmap

# parmap.map(process_document, remaining_paths_to_classify[-1000:], pm_processes=8, pm_pbar=True)

new_typed_dict = dict([(extract_filename(f), f) for f in getListOfFiles('/ivi/ilps/personal/vprovat/OCR_for_missing_folders/typed') if f.endswith('text.txt')])
new_handwritten_dict = dict([(extract_filename(f), f) for f in getListOfFiles('/ivi/ilps/personal/vprovat/OCR_for_missing_folders/likely_handwritten') if f.endswith('osd.txt')])
new_drawing_dict = dict([(extract_filename(f), f) for f in getListOfFiles('/ivi/ilps/personal/vprovat/OCR_for_missing_folders/likely_drawing')])

def path_already_processed(path):
    f = extract_filename(path)
    return (f in new_typed_dict) or (f in new_handwritten_dict) or (f in new_drawing_dict)

# in case we need to update
still_remaining_paths_to_classify = [path for path in remaining_paths_to_classify if not path_already_processed(path)
                                    ]

len(still_remaining_paths_to_classify)

import parmap
parmap.map(process_document, still_remaining_paths_to_classify, pm_processes=2, pm_pbar=True)

import pickle
pickle.dump(still_remaining_paths_to_classify, open('paths_to_classify.p','wb'))

def update_classification(row):
    if row['page_type']:
        return row['page_type']
    f = row['filename']
    if f in new_typed_dict:
        return "typed"
    if f in new_handwritten_dict:
        return "likely handwritten"
    if f in new_drawing_dict:
        return "likely drawing"
    
    return None

def update_osd(row):
    if row['path_to_handwritten_info']:
        return row['path_to_handwritten_info']
    
    f = row['filename']
    if f in new_handwritten_dict:
        return new_handwritten_dict[f]
    return None
    

def update_text(row):
    if row['path_to_text']:
        return row['path_to_text']
    
    f = row['filename']
    if f in new_typed_dict:
        return new_typed_dict[f]
    return None
    

df_files['page_type'] = df_files.apply(lambda row: update_classification(row),
                                       axis=1)

df_files['path_to_handwritten_info'] = df_files.apply(lambda row: update_osd(row),
                                       axis=1)

df_files['path_to_text'] = df_files.apply(lambda row: update_text(row),
                                       axis=1)

import csv

df_files.to_csv('archive_files_with_paths_descriptions_and_page_classifications.tsv',sep='\t', quoting=csv.QUOTE_NONNUMERIC)

df_files

# prepare for network analysis on letters: find letters, extract names (to, from)
# 

descriptions = sorted(list(set(df_files['part_of.title'].tolist())))

descriptions
