==> 05. Dataset stats.ipynb <==
import pandas as pd
df_entities = pd.read_csv('all_lettters_with_entities_final.tsv',sep='\t')


df_entities.head()

df_entities[df_entities['filename']==1288226]

df_entities.nunique()

import pandas as pd
df_graph = pd.read_csv('df_graph_with_wikidata.tsv',sep='\t')

df_graph

df_for_evaluation=pd.read_csv('sample_for_manual_evaluation_FINAL.tsv',sep='\t')

df_for_evaluation

wikidata = set(df_for_evaluation['wikidata_id1'].tolist()+df_for_evaluation['wikidata_id2'].tolist())

len(wikidata)
