==> 06.2. Build a graph based on manual annotations - sender-receiver only.ipynb <==
import pandas as pd
df_raw = pd.read_excel('Network_from_Carlotta/Correspondence_analysis.xlsx',sep='\t')

df = df_raw.dropna(subset=['Sender (From)','Addressee (To)'])

actors = set(df['Sender (From)'].tolist()).union(set(df['Addressee (To)'].tolist()))
actors

df

# import json
# alias = json.load(open('alias_dictionary_final.json','r'))


import json
aka = json.load(open('also_known_as_final.json','r'))

alias = {}
for main_name, names in aka.items():
    for name in names:
        alias[name] = main_name
        
def resolve_alias(p):
    cnt = 0
    while p in alias and cnt < 10:
        p = alias[p]
        cnt += 1
    if cnt == 10:
        print('Got a loop in aliases! Starting point: ',p)
    return p


graph = {}

def add_edge(p1, p2, doc_id):
    if p1 == p2:
        return
    
    if (p1, p2) in graph:
        graph[(p1,p2)].append(doc_id)
    else:
        graph[(p1,p2)] = [doc_id]
        
        
def add_edges(people, doc_id):
    '''
    people: list of people found in a document
    doc_id: path to text or another unique identifier
    '''
    if len(people) < 2:
        return
    
    for i, p1_raw in enumerate(people):
        for p2_raw in people[i:]:
            p1 = resolve_alias(p1_raw)
            p2 = resolve_alias(p2_raw)
            add_edge(p1, p2, doc_id)
#             add_edge(p2, p1, doc_id) # NB: directed graph in this case!

def extract_id(path):
    return path.split('/')[-1].split('.')[0]

from tqdm.auto import tqdm

def split_name(name): # for cases like Sybren and Veronique Valkema
    if name.find(' and ') == -1:
        return [name]
    
    first_part, second_part = name.split(' and ') # only two people together, not more
    
    if second_part.find(' ') != -1: # last name present
        if first_part.find(' ') == -1: # no last name for first person
            lastname = second_part.split(' ')[-1]
            first_part += ' '+lastname
    return [first_part,second_part]

for i in tqdm(range(0, len(df))):
    row = df.iloc[i]
    sender = row['Sender (From)'] 
    receiver = row['Addressee (To)']
    
    if sender.startswith('?') or receiver.startswith('?'):
        continue
    if sender.endswith('?') or receiver.endswith('?'):
        continue
    if sender == 'Sybren  Valkema':
        sender = 'Sybren Valkema'
    if receiver == 'Sybren  Valkema':
        receiver = 'Sybren Valkema'
        
    senders = split_name(sender) # for cases like Sybren and Veronique Valkema
    receivers = split_name(receiver) 
    filename = str(int(row['File name']))
    for s in senders:
        for r in receivers:
            add_edges([s, r],filename) 

# saving to a readable format

records = []
# seen_connections = set()
for people, docs in graph.items():
    person1 = min(people)
    person2 = max(people)
#     if (person1, person2) in seen_connections:
#         continue
#     seen_connections.add((person1,person2))
    records.append({'person1':person1,
                   'person2':person2,
                   'document_ids':docs})
    
df_graph = pd.DataFrame.from_records(records)
df_graph

df_graph = df_graph.groupby(['person1', 'person2'], as_index=False).agg({'document_ids': 'sum'})
df_graph

df_graph.to_csv('/ivi/ilps/personal/vprovat/valkema_graph_directed_manual_FINAL.tsv',sep='\t',index=False)

df_graph = pd.read_csv('/ivi/ilps/personal/vprovat/valkema_graph_directed_manual.tsv',sep='\t')

# from PIL import Image
# def show_image(filename):
#     df_target = df[df['filename'] == filename]
#     img_path_big = df_target.iloc[0]['path_to_image']
#     img_path = img_path_big.replace('TIF','JPG').replace('tif','jpg')
#     try:
#         img = Image.open(img_path)
#     except:
#         img = Image.open(img_path_big)
#     display(img)
#     print(img_path_big)

# show_image('0001246779.jpg')

import networkx as nx


import numpy as np

df_graph['weight']=df_graph['document_ids'].apply(lambda x: len(eval(x)))

# Old code begins her

%
def num_people(doc_id):
    return 2 # only doing sender-receiver for now
#     people = eval(df[df['filename']==doc_id+'.jpg'].iloc[0]['people'])
#     return len(people)

def get_weight(doc_ids):
    total = sum([num_people(doc_id) for doc_id in doc_ids.split('-')])
    return total # todo check it

def get_weight_for_list(doc_ids_list):
    return sum([1/get_weight(doc_id) for doc_id in doc_ids_list])

G = nx.Graph()
for item in graph.items():
    if len(set(item[1])) < 3: # pruning the long tail
        continue
    if item[0][0] == 'Valkema' or item[0][1] == 'Valkema': # ambiguous node, not useful
        continue 
    # (('M. Th. Brusse', 'Valkema'), ['0001246763'])
    weight = get_weight_for_list(item[1])
    G.add_edge(item[0][0], item[0][1], weight=weight)# np.log(len(set(item[1]))))

from pyvis.network import Network
def prepare_network(G1):
    net = Network(notebook=True,cdn_resources='remote')
    net.from_nx(G1)
    net.toggle_physics(True)
    return net

net = prepare_network(G)
net.repulsion()
net.show('graph_manual.html')

# now, Valkema-only

Valkema = nx.Graph()
for item in graph.items():
    if len(set(item[1])) < 3: # pruning the long tail
        continue
    if item[0][0] == 'Valkema' or item[0][1] == 'Valkema': # ambiguous node, not useful
        continue 
    if item[0][0] != 'Sybren Valkema' and item[0][1] != 'Sybren Valkema': # Valkema only
        continue 
    # (('M. Th. Brusse', 'Valkema'), ['0001246763'])
    weight = get_weight_for_list(item[1])
    Valkema.add_edge(item[0][0], item[0][1], weight=weight)# np.log(len(set(item[1]))))
    
net = prepare_network(Valkema)
net.repulsion()
net.show('Valkema.html')

# now, Valkema-only

Valkema = nx.Graph()
for item in graph.items():
    if len(set(item[1])) < 3: # pruning the long tail
        continue
    if item[0][0] == 'Valkema' or item[0][1] == 'Valkema': # ambiguous node, not useful
        continue 
    if not item[0][0].endswith('Valkema') and not item[0][1].endswith('Valkema'): # Valkema only
        continue 
    # (('M. Th. Brusse', 'Valkema'), ['0001246763'])
    weight = get_weight_for_list(item[1])
    Valkema.add_edge(item[0][0], item[0][1], weight=weight)# np.log(len(set(item[1]))))
    
net = prepare_network(Valkema)
net.repulsion()
net.show('Valkema.html')

# todo: ask Carlotta for her list of people, show her ours, discuss the graph
# ask how she came up with places - extracting from letters or knowing where everyone works?

# automatic network analysis for digital humanities
# RQs from Carlotta
# find centrality, PageRank etc
