==> 06.3. Prepare a smaller graph from the same docs as manual.ipynb <==
import pandas as pd
df_letters = pd.read_csv('all_lettters_with_entities_final.tsv',sep='\t')
df_manual = pd.read_excel('Network_from_Carlotta/Correspondence_analysis.xlsx',sep='\t').dropna(subset=['Sender (From)','Addressee (To)'])
filenames_manual = set(df_manual['File name'].tolist())
df_graph = pd.read_csv('df_graph_with_wikidata.tsv',sep='\t')
df_letters['is_seen_in_manual'] = df_letters['filename'].map(lambda x: x in filenames_manual)
df_letters_interesting = df_letters[df_letters['is_seen_in_manual']]

import json
aka = json.load(open('also_known_as_final.json','r'))

alias = {}
for main_name, names in aka.items():
    for name in names:
        alias[name] = main_name
        
def resolve_alias(p):
    cnt = 0
    while p in alias and cnt < 10:
        p = alias[p]
        cnt += 1
    if cnt == 10:
        print('Got a loop in aliases! Starting point: ',p)
    return p


people_with_wikidata = set(df_graph['person1'].tolist()+df_graph['person2'].tolist())
df_letters_interesting['people'] = df_letters_interesting['people_split'].map(lambda x: set([resolve_alias(person) for person in eval(x)
                                                            if resolve_alias(person) in people_with_wikidata]))
df_letters_interesting['has_people'] = df_letters_interesting['people'].map(lambda x: len(x)>1)
df_letters_interesting = df_letters_interesting[df_letters_interesting['has_people']].reset_index()
df_letters_interesting['filename'] = df_letters_interesting['filename'].astype(int)
df_letters_interesting.head()

graph = {}

def add_edge(p1, p2, doc_id):
    if p1 == p2:
        return
    
    if (p1, p2) in graph:
        graph[(p1,p2)].append(doc_id)
    else:
        graph[(p1,p2)] = [doc_id]
        
        
def add_edges(people, doc_id):
    '''
    people: list of people found in a document
    doc_id: path to text or another unique identifier
    '''
    if len(people) < 2:
        return
    
    for i, p1_raw in enumerate(people):
        for p2_raw in people[i:]:
            p1 = resolve_alias(p1_raw)
            p2 = resolve_alias(p2_raw)
            add_edge(p1, p2, doc_id)
            add_edge(p2, p1, doc_id)

def stops_here(cur_df):
    if cur_df.empty:
        return True
    try:
        text = open(cur_df.iloc[0]['path_to_text'],'r').read().lower()
    except:
        print('Failed to find text for: ',cur_df.iloc[0]['path_to_text'])
        text = ''
    if text.find('geachte') != -1:
        return True
    if text.find('geehrter') != -1:
        return True
    if text.find('dear') != -1:
        return True
    return False

def get_next_file(filename):
    name = filename+1
    return df_letters_interesting[df_letters_interesting['filename']==name]

from tqdm.auto import tqdm

processed_filenames = set()
for i in tqdm(range(0, len(df_letters_interesting))):
    row = df_letters_interesting.iloc[i]
    if row['filename'] in processed_filenames: # it was someone else's second page already
        continue 
    processed_filenames.add(row['filename'])

    cur_people = set(eval(row['people'])) if isinstance(row['people'],str) else row['people']

    cur_id = row['filename']
    
    df_next = get_next_file(row['filename'])
    while not stops_here(df_next):
        next_row = df_next.iloc[0]
        processed_filenames.add(next_row['filename'])

        next_people = set(eval(next_row['people'])) if isinstance(next_row['people'],str) else next_row['people']
        cur_people = cur_people.union(next_people)

        cur_id += '-'+ extract_id(next_row['path_to_image'])
        
        df_next = get_next_file(next_row['filename'])
        
    add_edges(list(cur_people), cur_id)

wikidata_id_per_name = dict(zip(df_graph['person1'].tolist()+df_graph['person2'].tolist(), 
                                df_graph['wikidata_id1'].tolist()+df_graph['wikidata_id2'].tolist()
                               ))

wikidata_description_per_name = dict(zip(df_graph['person1'].tolist()+df_graph['person2'].tolist(), 
                                df_graph['wikidata_description1'].tolist()+df_graph['wikidata_description2'].tolist()
                               ))

main_country_per_name = dict(zip(df_graph['person1'].tolist()+df_graph['person2'].tolist(), 
                                df_graph['main_country1'].tolist()+df_graph['main_country2'].tolist()
                               ))

# saving to a readable format

records = []
seen_connections = set()
for people, docs in graph.items():
    person1 = min(people)
    person2 = max(people)
    if (person1, person2) in seen_connections:
        continue
    seen_connections.add((person1,person2))
    records.append({'person1':person1,
                   'person2':person2,
                   'document_ids':docs,
                   'weight':len(docs),
                   'wikidata_id1':wikidata_id_per_name[person1],
                   'wikidata_id2':wikidata_id_per_name[person2],
                   'wikidata_description1':wikidata_description_per_name[person1],
                   'wikidata_description2':wikidata_description_per_name[person2],
                   'main_country1': main_country_per_name[person1],
                   'main_country2': main_country_per_name[person2]})
    
df_graph_small = pd.DataFrame.from_records(records)
df_graph_small

df_graph_small.to_csv('df_graph_same_docs_as_manual.tsv',sep='\t',index=False)
