==> 10. Graph stats.ipynb <==
import pandas as pd
df_graph = pd.read_csv('df_graph_with_wikidata.tsv',sep='\t')
df_graph.head()

import networkx as nx
df_graph_common = df_graph[df_graph['weight']>1]

G = nx.Graph()
for _,row in df_graph.iterrows():
    G.add_edge(row['person1'],row['person2'],weight=row['weight'])

import networkx as nx
from sklearn.preprocessing import MinMaxScaler

def top_n_centrality(G,N=10):
    # Calculate centralities
    degree_centrality = nx.degree_centrality(G)
    weighted_degree = dict(G.degree(weight='weight')) 
    betweenness_centrality = nx.betweenness_centrality(G)# not adding weight here because it's interpreted as distance
    closeness_centrality = nx.closeness_centrality(G)
    largest_component_subgraph = G.subgraph(max(nx.connected_components(G),key=len)).copy()
    pagerank = pnx.pagerank(largest_component_subgraph)

    
    # Adjust max_iter and tol to help with convergence
    try:
        eigenvector_centrality = nx.eigenvector_centrality(G, weight='weight', max_iter=1000, tol=1e-4)
    except nx.PowerIterationFailedConvergence:
        print("Eigenvector centrality did not converge.")


    # Create a DataFrame for centrality metrics
    centrality_df = pd.DataFrame({
        'degree': degree_centrality,
        'weighted_degree':weighted_degree,
        'betweenness': betweenness_centrality,
        'closeness': closeness_centrality,
        'eigenvector': eigenvector_centrality,
        'pagerank':pagerank,
    })

    # Normalize centralities using MinMaxScaler
    scaler = MinMaxScaler()
    normalized_df = pd.DataFrame(scaler.fit_transform(centrality_df), columns=centrality_df.columns, index=centrality_df.index)

    # Compute a composite score by summing the normalized centralities
    normalized_df['composite'] = normalized_df.sum(axis=1)

    # Sort nodes by the composite score
    sorted_nodes = normalized_df.sort_values(by='composite', ascending=False).round(3)


    return sorted_nodes[:N]

top_n_centrality(G,10)

# Now, the manual graph for comparison

df_graph_manual = pd.read_csv('/ivi/ilps/personal/vprovat/valkema_graph_directed_manual_FINAL.tsv',sep='\t')

df_graph_manual['weight']=df_graph_manual['document_ids'].apply(lambda x: len(eval(x)))
df_graph_manual

# def split_name(name): # for cases like Sybren and Veronique Valkema
#     if name.find(' and ') == -1:
#         return [name]
    
#     first_part, second_part = name.split(' and ') # only two people together, not more
    
#     if second_part.find(' ') != -1: # last name present
#         if first_part.find(' ') == -1: # no last name for first person
#             lastname = second_part.split(' ')[-1]
#             first_part += ' '+lastname
#     return [first_part,second_part]

# split_name('Sybren and Veronique Valkema')

G_manual = nx.Graph()

for _,row in df_graph_manual.iterrows():
    person1, person2 = row['person1'],row['person2']
    
#     if person1.startswith('?') or person2.startswith('?'):
#         continue
#     if person2 == 'Sybren  Valkema':
#         person2 = 'Sybren Valkema'
        
    weight=row['weight']
    if weight > 1:
#     people1, people2 = split_name(person1),split_name(person2)
#     for per1 in people1:
#         for per2 in people2:
        G_manual.add_edge(person1, person2,weight=weight)

top_n_centrality(G_manual,10)

# Now, the graph created automatically on the same sample as the manual grapjh

df_graph_small = pd.read_csv('df_graph_same_docs_as_manual.tsv',sep='\t')

G_small = nx.Graph()

for _,row in df_graph_small.iterrows():
    person1, person2 = row['person1'],row['person2']
    
        
    weight=row['weight']
    if weight > 1:
#     people1, people2 = split_name(person1),split_name(person2)
#     for per1 in people1:
#         for per2 in people2:
        G_small.add_edge(person1, person2,weight=weight)

top_n_centrality(G_small,10)

%
import sys
!{sys.executable} -m pip uninstall community -y
!{sys.executable} -m pip uninstall python-louvain -y
!{sys.executable} -m pip install python-louvain

# Now, graph-wide statistics for everyone

import community.community_louvain as community_louvain

def get_graph_metrics(G_raw,graph_name):
    partition = community_louvain.best_partition(G_raw)
    G = G_raw.subgraph(max(nx.connected_components(G_raw),key=len)).copy() # if the graph is not connected, work with largest component
    return {
        'graph_name':graph_name,
        'num_components': nx.number_connected_components(G_raw),
        'num_nodes': G.number_of_nodes(),
        'avg_degree': sum(dict(G.degree()).values()) / G.number_of_nodes(),
        'avg_weighted_degree': sum(dict(G.degree(weight='weight')).values()) / G.number_of_nodes(),
        'density': nx.density(G),
        'diameter': nx.diameter(G),
        'avg_clustering': nx.average_clustering(G),
#         'partition': partition,
        'modularity': community_louvain.modularity(partition, G)
    }
        
records=[get_graph_metrics(G,'graph'),
         get_graph_metrics(G_manual,'graph_manual'),
         get_graph_metrics(G_small,'graph_small')]

df_graph_stats = pd.DataFrame.from_records(records)
df_graph_stats.round(2)

pagerank = nx.pagerank(G)
print(f"PageRank: {pagerank}")


# Old code begins here

def metrics_to_df(metrics,metric_name='degree_centrality'):
    return pd.DataFrame.from_records(sorted(metrics.items(), key=lambda item: item[1], reverse=True),
                                                    columns=['name',metric_name
                                                            ]).round(3)

degree_centrality_df = metrics_to_df(nx.degree_centrality(G))
betweenness_centrality_df = metrics_to_df(nx.betweenness_centrality(G),'betweenness_centrality')
closeness_centrality_df = metrics_to_df(nx.closeness_centrality(G),'closeness_centrality')
eigenvector_centrality_df = metrics_to_df(nx.eigenvector_centrality(G),'eigenvector_centrality')

eigenvector_centrality_df

sorted(degree_centrality.items(), key=lambda item: item[1], reverse=True)
